Add re/decommission toolbard buttons
In order for the two operations to exist in the UI we need to add them
to the UI. We have chosen a dedicated toolbar button group for Cisco
Intersight provider as there will be some more in the future. These
operations do not make much sense for other providers, though.