Set common event type for alarms
Alarms in Intersight can represent various information to the user. In
order to preserve the relevant information, we want to store all
alarms in the event stream and expose them in the timeline. To this
end we are setting the `event_type` to a common `Alarm` type and
use the `name` of the event in the message. This way, manageiq users
will still see all the relevant information as it occurs.